Crate namada_merkle_tree

The merkle tree in the storage

Modules

eth_bridge_pool

Ethereum bridge pool merkle tree

ics23_specs

A module that contains

Structs

BlockHeight

Height of a block, i.e. the level. The default is the BlockHeight::sentinel value, which doesn’t correspond to any block.

CommitDataRoot

Fake merkle tree just to commit extra data to the merkle tree

CommitmentProof

CommitmentProof is either an ExistenceProof or a NonExistenceProof, or a Batch of such messages

Epoch

Epoch identifier. Epochs are identified by consecutive numbers.

H256

Represent 256 bits

Hash

A hash, typically a sha-2 hash of a tx

KeccakHash

Represents a Keccak hash.

Key

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

MerkleRoot

The root hash of the merkle tree as bytes

MerkleTree

Merkle tree storage

MerkleTreeStoresRead

The root and store pairs to restore the trees

MerkleTreeStoresWrite

The root and store pairs to be persistent

Proof

A storage key existence or non-existence proof

Enums

Error

MembershipProof

Type of membership proof from a merkle tree

Store

Backing storage for merkle trees

StoreRef

Pointer to backing storage of merkle tree

StoreType

Store types for the merkle tree

Constants

NO_DIFF_KEY_PREFIX

Key prefix for the data not stored to diffs

Traits

StorageHasher

The storage hasher used for the merkle tree.

SubTreeRead

Trait for reading from a merkle tree that is a sub-tree of the global merkle tree.

SubTreeWrite

Trait for updating a merkle tree that is a sub-tree of the global merkle tree

Functions

tree_key_prefix_with_epoch

Get the key prefix with which the tree root and store are stored in the storage

tree_key_prefix_with_height

Get the key prefix with which the tree root and store are stored in the storage

Type Aliases

Amt

Arse-merkle-tree

AmtStore

Arse-merkle-tree store

BridgePoolStore

Bridge pool store

Smt

Sparse-merkle-tree

SmtStore

Sparse-merkle-tree store

StorageBytes

Type alias for bytes to be put into the Merkle storage